+++ src/nodeTreeDynamicWidgets/memoryObjectDynamicTreeWidgetItem.cpp
@@ -19,9 +19,10 @@
 
 #include "memoryObjectDynamicTreeWidgetItem.hpp"
 
-MemoryObjectDynamicTreeWidgetItem::MemoryObjectDynamicTreeWidgetItem(la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::MemoryObjectIndex const memoryObjectIndex, la::avdecc::controller::model::MemoryObjectNodeDynamicModel const* const dynamicModel, QTreeWidget *parent)
+MemoryObjectDynamicTreeWidgetItem::MemoryObjectDynamicTreeWidgetItem(la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::ConfigurationIndex const configurationIndex, la::avdecc::entity::model::MemoryObjectIndex const memoryObjectIndex, la::avdecc::controller::model::MemoryObjectNodeDynamicModel const* const dynamicModel, QTreeWidget *parent)
 	: QTreeWidgetItem(parent)
 	, _entityID(entityID)
+	, _configurationIndex(configurationIndex)
 	, _memoryObjectIndex(memoryObjectIndex)
 {
 	// MemoryObjectLength
@@ -34,9 +35,9 @@ MemoryObjectDynamicTreeWidgetItem::MemoryObjectDynamicTreeWidgetItem(la::avdecc:
 		updateMemoryObjectLength(dynamicModel->length);
 
 		// Listen for MemoryObjectLengthChanged
-		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::memoryObjectLengthChanged, this, [this](la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::MemoryObjectIndex const memoryObjectIndex, std::uint64_t const length)
+		connect(&avdecc::ControllerManager::getInstance(), &avdecc::ControllerManager::memoryObjectLengthChanged, this, [this](la::avdecc::UniqueIdentifier const entityID, la::avdecc::entity::model::ConfigurationIndex const configurationIndex, la::avdecc::entity::model::MemoryObjectIndex const memoryObjectIndex, std::uint64_t const length)
 		{
-			if (entityID == _entityID && memoryObjectIndex == _memoryObjectIndex)
+			if (entityID == _entityID && configurationIndex == _configurationIndex  && memoryObjectIndex == _memoryObjectIndex)
 			{
 				updateMemoryObjectLength(length);
 			}